+++ docs/Energinet.DataHub.MarketParticipant.Authorization/release-notes/release-notes.md
@@ -1,5 +1,23 @@
 # Energinet.DataHub.MarketParticipant.Authorization release notes
 
+## Version 1.1.1
+
+- Improve IEndpointAuthorizationLogger registration.
+
+## Version 1.1.0
+
+- Added IEndpointAuthorizationContext.
+- Added logging support through IEndpointAuthorizationLogger.
+- Added RequestSignatureAsync with User Id.
+
+## Version 1.0.2
+
+- Added energysupplier role validation for access validation.
+
+## Version 1.0.1
+
+- Added support for key rotation.
+
 ## Version 1.0.0
 
 - First implementation of Authorization